+++ compat-table/src/index.ts
@@ -89,6 +89,7 @@ export const jsFeatures = {
 
 export type CSSFeature = keyof typeof cssFeatures
 export const cssFeatures = {
+  ColorFunction: true,
   HexRGBA: true,
   HWB: true,
   InlineStyle: true,