+++ internal/css_parser/css_decls.go
@@ -151,8 +151,19 @@ func (p *parser) processDeclarations(rules []css_ast.Rule, composesContext *comp
 			}
 
 		case css_ast.DBackground:
+			wouldClamp := false
 			for i, t := range decl.Value {
-				decl.Value[i] = p.lowerAndMinifyColor(t)
+				decl.Value[i] = p.lowerAndMinifyColor(t, &wouldClamp)
+			}
+
+			// If this contains values outside of sRGB, duplicate and clamp
+			if wouldClamp && p.options.unsupportedCSSFeatures.Has(compat.ColorFunction) {
+				clamped := *decl
+				clamped.Value = css_ast.CloneTokensWithoutImportRecords(clamped.Value)
+				for i, t := range clamped.Value {
+					clamped.Value[i] = p.lowerAndMinifyColor(t, nil)
+				}
+				rewrittenRules = append(rewrittenRules[:len(rewrittenRules)-1], css_ast.Rule{Loc: rule.Loc, Data: &clamped}, rule)
 			}
 
 		case css_ast.DBackgroundColor,
@@ -178,7 +189,16 @@ func (p *parser) processDeclarations(rules []css_ast.Rule, composesContext *comp
 			css_ast.DTextEmphasisColor:
 
 			if len(decl.Value) == 1 {
-				decl.Value[0] = p.lowerAndMinifyColor(decl.Value[0])
+				wouldClamp := false
+				decl.Value[0] = p.lowerAndMinifyColor(decl.Value[0], &wouldClamp)
+
+				// If this contains values outside of sRGB, duplicate and clamp
+				if wouldClamp && p.options.unsupportedCSSFeatures.Has(compat.ColorFunction) && len(decl.Value) == 1 {
+					clamped := *decl
+					clamped.Value = css_ast.CloneTokensWithoutImportRecords(clamped.Value)
+					clamped.Value[0] = p.lowerAndMinifyColor(clamped.Value[0], nil)
+					rewrittenRules = append(rewrittenRules[:len(rewrittenRules)-1], css_ast.Rule{Loc: rule.Loc, Data: &clamped}, rule)
+				}
 			}
 
 		case css_ast.DTransform:
@@ -187,8 +207,15 @@ func (p *parser) processDeclarations(rules []css_ast.Rule, composesContext *comp
 			}
 
 		case css_ast.DBoxShadow:
-			if p.options.minifySyntax {
-				decl.Value = p.mangleBoxShadows(decl.Value)
+			wouldClamp := false
+			decl.Value = p.lowerAndMangleBoxShadows(decl.Value, &wouldClamp)
+
+			// If this contains values outside of sRGB, duplicate and clamp
+			if wouldClamp && p.options.unsupportedCSSFeatures.Has(compat.ColorFunction) {
+				clamped := *decl
+				clamped.Value = css_ast.CloneTokensWithoutImportRecords(clamped.Value)
+				clamped.Value = p.lowerAndMangleBoxShadows(clamped.Value, nil)
+				rewrittenRules = append(rewrittenRules[:len(rewrittenRules)-1], css_ast.Rule{Loc: rule.Loc, Data: &clamped}, rule)
 			}
 
 		// Container name